commented code

+++ src/Food.h
@@ -5,17 +5,21 @@
 #include <stdlib.h>
 #include <sys/time.h>
 
+/* Structs */
+/* Struct holding a single food item */
 typedef struct Bite {
 	uint16_t x;
 	uint16_t y;
 } Bite;
-
+/* Struct containing all food items on screen or game */
 typedef struct Food {
 	Bite *arr;
 	uint8_t size;
 	WINDOW *win; 
 } Food;
 
+/* Initializes a Food struct with given window and size.
+ */
 Food initFood(WINDOW* win, uint8_t size)
 {
 	Food f = {
@@ -26,11 +30,13 @@ Food initFood(WINDOW* win, uint8_t size)
 	return f;
 }
 
+/* If food array is created using initializer or malloc, this method frees used memory. */
 void freeFood(Food* f)
 {
 	free(f->arr);
 }
 
+/* Sets the x and y coordinates of a specific food item. */
 void setFood(Food *f, uint8_t index, uint16_t x, uint16_t y)
 {
 	if(index >= 0 && index < f->size)
@@ -40,6 +46,9 @@ void setFood(Food *f, uint8_t index, uint16_t x, uint16_t y)
 	}
 }
 
+/* Sets the x and y coordinates of a specific food item to random values. 
+ * Is seeded by timeval.tv_usec because must be used multiple times in rapid succession 
+ */
 void setFoodRand(Food *f, uint8_t index)
 {
 	struct timeval tv;
@@ -54,19 +63,24 @@ void setFoodRand(Food *f, uint8_t index)
 	}
 }
 
+/* Remove all food items from screen */
 void eraseFood(Food *f)
 {
 	for(int i = 0; i < f->size; i++)
 		mvwaddch(f->win, f->arr[i].y, f->arr[i].x, ' ');
 }
 
+/* Draw all food items on screen */
 void drawFood(Food *f)
 {
 	for(int i = 0; i < f->size; i++)
 		mvwaddch(f->win, f->arr[i].y, f->arr[i].x, 'O');
 }
 
-int16_t isIn(Food *f, uint16_t x, uint16_t y)
+/* Returns whether  or not a coordinate pair is the same as a food item's position. 
+ * Used to check for collisions
+ */
+int8_t isIn(Food *f, uint16_t x, uint16_t y)
 {
 	for(uint8_t i = 0; i < f->size; i++)
 	{
